+++ changelog.md
@@ -1,11 +1,15 @@
-## Changes and new features
+### Changes and new features
 
 - Added per-map option for materials counting;
 - Moving/scaling an image with mouse in manual cropping became more accurate;
+- Added lifetime for closest colors cache that can significantly speed up image converting process when making a few changes in a row;
 
-### Fixes and minor changes
+<details>
+<summary>Fixes and minor changes</summary>
 
 - Slightly improved positioning of fake map position for even maparts;
 - Improved `/mart-creative give-map-palette` - now it creates a new locked map;
 - Fixed some exceptions when changing mapart size during converting;
 - Mainly, this update contains codebase changes and reworks which are invisible for users
+
+</details>